backend/requirements.txt
fastapi
uvicorn[standard]
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-qdrant>=0.1.0
qdrant-client>=1.8.0
psycopg2-binary
python-dotenv
pydantic
pytest